// packages/request-builder/src/types.ts
export interface Builder {
  build: (...args: any[]) => AjaxConfig;
}

enum HttpMethods {
  GET = 'get',
  PUT = 'put',
  POST = 'post',
  DELETE = 'delete',
  OPTIONS = 'options',
  HEAD = 'head',
  PATCH = 'patch',
  // TRACE = 'trace'
}

export const METHODS = Object.values(HttpMethods);

export type RequestParams = Partial<{ params: Record<string, any>; body: any; }>;